Add OpenAPI definition for 'mini-manage' contract
A first attempt at defining the contract between the UI app
and this API app.

The process we're planning is:

1. UI app stubs out these endpoints using local fixtures in
its app

2. This API app implements these endpoints and stubs out the
responses using local fixture here in this app. The OpenAPI
spec can be replaced by annotations in the Kotlin
controllers which will be built, e.g.:
  - @Schema
  - @Operation
  - @PathVariable
  - @RequestBody
  - @ApiResponses
  - @PostMapping

3. This API app makes the necessary upstream connections to
HMPPS services (Community API, ARN, Prisons etc) and builds
the representations described in the OpenAPI spec.