Add scripts to rule them all
This adds some simple script to aid running in development. We use these
in the frontend application, so it’s useful to mirror this approach in
the API too, to aid with context switch. Currently the bootsrap and
server scripts are stubs, as there’s nothing we (yet) need to do to
bootstrap / setup the app, but they’re worth having around as a reminder
to update them when we do start introducing external dependencies, like
a database etc.