Return more detailed problem responses when incorrect request bodies are provided
Jackson (the deserializer used by Spring) fails fast when deserializing into POJO, so it's only possible out of the box to provide the first error encountered during deserialization back to the caller.  This catches those exceptions then uses reflection to check each property and provide more detailed responses.  For this to work, we need to read the input stream more than once, so this also uses ContentCachingRequestWrapper to read the input stream once and cache it for reuse.