OpenAPI changes for Key Workers
Remove keyWorkerId from NewBooking as this will now be provided on Arrival.
Add keyWorkerStaffId to NewArrival
Rename the KeyWorker component to StaffMember as all Key Workers are Staff but not vice-versa.  We will be returning all Key Workers for a team to be selected as a Key Worker.
Remove isActive from KeyWorker (StaffMember) component as this is not a field exposed by Community API.