Add person detail to Application responses
This adds the detail of a specific person to all of the Application
responses. This involves us having to call the Community and Prisons
APIs, so we have guards in place for if either API has problems
returning the person detail(s).

This is done using a private method in the controller, the
implementation of which is nicked from the BookingsController. I do
wonder if this would be better placed in a service, but I think this is
fine for now.

I’ve also added the `mockInmateDetailPrisonsApiCall` function, which is
used in some other integration tests to the IntegrationTestBase file, so
we can use it everywhere.