Redis uses a different port in development to fix conflict with UI
When both applications start up they expected to both use the default port and be accessible at localhost:6379. Two containers can't be bound to the same port so one of the apps has to change from the default.

This change makes the API redis container available at localhost:6380, allowing the frontend to use the default.