Remove docker container names to prevent naming confclits
Setting names can conflict with other projects, such as the frontend.

When we remove the name, docker compose will use the directory name as a prefix, ensuring uniqueness: `hmpps-approved-premises-ui_wiremock_1` rather than `hmpps-wiremock_1`.

This removes a conflict with wiremock with the frontend and allows both projects to be started independently.