Add deliusEventNumber to Domain Events
The Delius event number is a unique code which, in tandem with the CRN,
identifies a person's current "event" within Delius. "Event" is a core
concept within Delius and maps broadly to conviction.

The AP service is going to need to retrieve the `DeliusEventNumber` from
Community API, probably on the "Confirm identity" step. This
`deliusEventNumber` looks like it will be a "first-class" property of the
"application" entity, e.g. via a db property such as `applications.index_offence`
JSON blob consisting of:

- `convictionId`
- `eventNumber`
- `offenceId`

The Delius event number is the `index` property of the first listed
conviction, i.e.

```
curl https://community-api/secure/offenders/crn/{crn}/convictions \
  | jq '[.[] | select(.active)][0] | .index'
```
and will be a number in string form, e.g. `"7"`.

Further to discussion with the Delius team we've added this property to
all Domain Events.