Remove submission capability from PUT application endpoint
Since we are splitting out a submission endpoint, we no longer need to do this from the PUT endpoint when a submittedAt value is provided.

Do not validate schema on PUT endpoint

Since we cannot reliably validate against the schema until the entire application has been completed, we should not do this from the PUT endpoint.