Add applicationId to booking and person events
A stay in an AP needs to be identified with a particular “application” as
the application (and its assessment) constitute a risk assessment of
AP-specific risk at a point in time. If there are is an incident whilst
residing at an AP it’s important that there is a clear link to the application
which led to the placement.

In some cases a person will stay at multiple APs (same or different premises)
under the authority of a single application (which we think will be valid
for 9 months) e.g:

- regular temporary releases via ROTL (release on temporary licence)
- release → recall → re-release cycle

It's our understanding that the application ID will be used within Delius
when creating "contacts" and potentially updating the Residence NSI in
the course of processing events.