Add applicationUrl property to Domain Events
It's important that Domain Events provide a link to the UI of the
Approved Premises service, to a resource which will allow a human being to
explore the application and its associated resources e.g.:

- the assessment of the application (including clarification notes)
- any bookings related to the application

For example, when a `booking.made` event is consumed by the Delius integration
we expect a Delius "contact" to be created. This contact will be of the
new (booking) type `EAPB` and will include the `applicationUrl` e.g.

https://approved-premises/applications/484b8b5e-6c3b-4400-b200-425bbe410713

Probation officers using Delius will be able to follow this link and view
the (possibly multiple) bookings which have been made and explore the AP
"application" in full according to their needs.