Add OpenAPI spec for Documents download endpoint
There is unfortunately a compromise here in terms of the URL.  Logically `/applications/{applicationId}/documents/{documentId}` would have been the choice.  However, we do not want to use the controller that OpenAPI Generate creates because it doesn't support proxying to a file download in an upstream service in a memory efficient way.  Consequently, we have used `/documents/{crn}/{documentId}` so that OpenAPI Generate places the endpoint's function in a separate Controller class which can then be deleted from the output directory.  This way we still have the download endpoint in the api.yml/Swagger but are free to implement it in a more efficient way.