Implement download document endpoint
This first checks that the requested ID exists in the list Document meta-data response for the CRN and returns a 404 if not.  If the file is found, its contents is streamed back to the caller.  The streaming is achieved by combining StreamingResponseBody (https://www.baeldung.com/spring-mvc-sse-streams#streamingresponsebody) which lets us write directly to the output stream with DataBuffer (https://manhtai.github.io/posts/flux-databuffer-to-inputstream/) on the upstream Client which allows us to read directly from the stream.