Scaffolding for running seed jobs
Each type of Job must implement the SeedJob<RowType> interface.  Instances of these Job types are created by the SeedService which accepts a type of job and a filename.  The filename is loaded from a location on disk specified by the `seed.file-prefix` property which will be `/seed` on our Kubernetes deployment.