Change target location example to postcode district rather than place name
At present this is described as a PDU location given with the first part
of a postcode ([Postcode District](https://www.ons.gov.uk/methodology/geography/ukgeographies/postalgeography)) e.g. `LS2` for Woodhouse in Leeds or `BD1` for central Bradford. In the future we expect the service to record a target location as a point to facilitate placement as close as possible to the optimal resettlement location for the person.

It was suggested that Local Authority Area (e.g. `Leeds`, `Kirklees`)
might be a good alternative, as this fits more naturally with the CAS3
landscape where accommodation is tracked by local authority.

i.e. in the immediate term we plan to report `targetLocation` with the
post code area (e.g `LS2` or `BD1`)