Enable log-client-credentials-jwt-info on preprod
This setting causes the app to log the client_id, authorities, scopes and expiry when it retrieves a client_credentials JWT for service->service call authentication.  Turing this on in pre-prod to debug an issue with calling Offender Case Notes Service (we need to ensure we have the correct authorities.)  This does not log the JWT signature so is safe to enable.