Add applicationSubmittedOn to person.arrived event
It's been agreed the Residence NSI will be created in Delius when the
`<person.arrived>` event is consumed. The referral date of the NSI should
be the date at which the referral was made (i.e. `<application.submitted>.submittedAt`).

We could take a purist view that `person.arrived` should only describe
the arrival and perhaps offer a link back to the original `<application.submitted>`
event where the consumer can find the `submittedAt` datestamp.

Instead we take a pragmatic approach and provide an `applicationSubmittedOn`
date here as part of `<person.arrived>`.