Fix failing 'create non-arrival' test
As dates are randomly chosen, occasionally the booking would be given a date
before the non-arrival date used in the test, which would cause a 400 Bad
Request response with the error message 'afterBookingArrivalDate'.

This test has been changed to use the arrival date of the booking as the date
for the non-arrival.