Expose list of users to ROLE_ADMIN only
The list is sorted by name.

Our plan is that first phase of the Role Admin area will comprise a list
of users and their role. Adjustments to the role assignment will be managed
through the existing user seeding mechanism. `See [UsersSeedJob](https://github.com/ministryofjustice/approved-premises-api/blob/expose-users-and-roles-to-role-admin/src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/seed/UsersSeedJob.kt`)

The facility to add and remove roles through the UI will follow.

In due course we'll need to paginate this list. For the initial release in the
we expect about 250 user with roles to be managed.

Perhaps we'll want some filtering, e.g.

- only show users with internally managed roles
- show users with a particular role