Allow risks to be skipped with a flag set
This allows us to set a flag to skip the creation of risks when
creating an application. This will allow us to go live without
the reliance on the OASys integration, which may not be
ready for beta. If we do use this, we’ll need to backfill any
applications without risks added, which may be a manual
process.