Cope with missing email property from Community API StaffMember
When seeding users and roles we discovered that many of our Managers,
Assessors and Workflow Managers do not have an email in their Delius
`StaffMember` entity as represented via the `api/secure/staff/username/`
endpoint.

This commit makes the email property optional.

We need to liaise with the F/E team to ensure that this doesn't introduce
a breaking change where that `email` attribute is displayed to an Assessor
who wishes to ask the Applicant a clarification question. See:

https://github.com/ministryofjustice/hmpps-approved-premises-ui/pull/500