Update PersistedFactory to use new factory instance per call
This changes PersistedFactory to take a producer function for the factory rather than a single instance.  This allows us to use a fresh factory for each call to `produceAndPersist` which in turn prevents hard to debug issues caused by earlier calls to the factory setting values to literals (e.g. two Users being produced with the same deliusUsername because the first call explicitly sets it).