Add feature flag for assigning a default probation region to users
In non-production-like environments, the Community API returns probation area codes
that are not 1:1 with probation regions. To avoid polluting our probation regions
table and to avoid redeploying every time we want to add a new user to the dev/test
regions, the environment can be configured with:

```
assign-default-region-to-users-with-unknown-region: true
```

To allow a default value of 'North West' to be given as a probation region to users
whose probation area code in the Community API doesn't correspond to one of our
probation regions.