Users can be filtered by role and qualification
This adds a new JPA Specification, which accepts a list of
roles and qualifications, and optionally builds up a list of
predicates, depending on if the roles/qualifications are provided.

We can then provide the specification to the service, which also
converts the API roles / qualifications to the JPA enums.

I’ve also tidied up the tests, using the `Given a User` helpers, to
save us having to do a ton of setup/teardown in each test.