Changes to Person Departed in Domain Events Open API spec
Departure Reason and Legacy Reason code are both driven from the database so can't be specified as enums in the OpenAPI spec.
Move on Category Description and Legacy Category code are also driven from the database so can't be enumerated in spec.
Move on Category and Destination Provider should be required.

Change `code` to `id` on Move on Category & Destination Provider
These are our concepts so have a UUID id rather than a code.