Seed coordinates in Premises seeding
We add `latitude` and `longitude` columns to the generic `premises` table,
using the "precision double" type with 15 decimal digits precision.

https://www.postgresql.org/docs/current/datatype-numeric.html#DATATYPE-NUMERIC-DECIMAL

We expect that once we begin to do geospatial queries using this information
we'll migrate both coordinates to a PostGIS `geometry(point, 4326)` data type.

We've added the latitude and longitude columns to the shared premises table
as we believe that geospatial information is common and core to all varieties of
accommodation.