Set characteristics from individual columns
Rather than from a comma-delimited list in a single `characteristics`
column.

We now verify that each given characteristic is scoped to both:

- the service "approved-premises", and
- the model "premises"

from the `processRow()` function rather than from `deserializeRow()`.
This is because building the list of "positive" characteristics (those with
a YES value) is a little different and now uses a `CharacteristicValue`
data class to map the `propertyName`s and `value`s.