Reallocation endpoint supports multiple tasks
This updates the reallocate endpoint to support multiple tasks. I’ve
updated the body to accept a NewRellocation object with a Task Type and
user ID. Rather than returning the Assessment object (which we don’t
need), the endpoint now returns a Reallocation object with the Task Type
and the user who the task has been reallocated to.

As we don’t currently support any task type other than assessements, we
return a 405 method not allowed response. I’ve also added some tests for
all the other types to serve as placeholders, which can be replaced as
we add support for the other task types.