Add service to create/allocate a Placement Request
This adds a new service method to create a placement request in the
database when given an assessment and a PlacementRequest API object.

We also convert the enums to database objects by querying by the
property name with an IN query. For now, we trust that these will be in
the database, as they’re all seeded. In the future, we may tweak this to
accept UUIDs, but this adds additional complexity at the assessment
side, so we’re accepting enums for now.

We also allocate a request to a random user. I considered doing
users with the least pending requests, but random seems fairer.
We can iterate this as we go, and requests can always be
reallocated.