Add an endpoint to get an assessment as a task
This fetches an Assessment for an application, as returns it as a task.

I’ve also added a new Given helper, to try and tidy up some of the
boilerplate when creating assessments.

As with allocations, I’ve decided to return a 405 if the task type is
not yet implemented.