Include a load test for the summary endpoint
Given the nature of this test is to check how the service performs when lots of data is present it is intentionally slow. It creates 200 premises which takes time. We want this confidence but we don't want the cost of running it on every test run. For now it's excluded from the default test suite.

In future we could add this in as a parallel CI test.

This borrows from the first PR that was addressing n+1 queries [1].

[1] https://github.com/ministryofjustice/hmpps-approved-premises-api/pull/493