Fix NonUniqueResultException when checking for conflicts with bookings/lost beds
The original logic (written in Kotlin) used `.firstOrNull()` to get a single booking or lost bed that conflicts (if any existed).
The logic was later converted into JPQL queries which returned a nullable single result, which works as long as the query returns 0 or 1 results,
but more than one results in a NonUniqueResultException.
Unfortunately, JPQL doesn't have any `LIMIT 1`/`FETCH FIRST ROW ONLY` construct, so the repository methods that execute the JPQL queries have been updated to return a list.
The service methods then call `.firstOrNull()` on the list returned by the repository to achieve the same result as the original logic.