Ensure requirements are copied when reallocated
Before, when a placement request was reallocated, the requirements were
not copied over. To make this work properly, we need to load the
relation by calling `toList` on each collection. We can’t eagerly load
the relation as we get a `MultipleBagsException`. There are all
manner of complex workarounds, but this seems to do the trick and
only results in two additional queries, so not worth the additional stress
and complexity.