Implement support for either PDU name or ID when creating or updating a Temporary Accommodation premises
To maintain backwards compatibility while transitioning to modelling premises PDUs in a structured way, when creating or updating
a premises, the `PremisesService` now accepts an `Either<String, UUID>` that represents both the legacy PDU name as well as the
ID of a PDU that exists in the database.
The `PremisesService.tryGetProbationDeliveryUnit` method is used to perform the validation for the legacy `pdu` name field or
the `probationDeliveryUnitId` UUID (as appropriate) and using the relevant repository method to obtain a `ProbationDeliveryUnitEntity`
instance.

From this point, the use of `probationDeliveryUnitId` is preferred, which is reflected in the use of `UUID` as the second type argument
of `Either` (which is right-biased) and the validation error `"empty"` being reported on that property instead of `pdu` when neither
has a value.

The PDU is still stored in the premises as a string name rather than as an ID to a foreign key table, but performing the work to resolve
the name or ID now will reduce the work needed to migrate premises away from string PDU names.