Fix for Wiremock port race condition when running tests
Stopping Wiremock in the @AfterEach gives a period of time before it is restarted in @BeforeEach in which another test runner process can claim the port using find available.  Instead, use nio file locks support to get a lock on a temporary file representing a port in the range 57830->57880.  This allows us to start/stop the Wiremock server wherever without worrying about race conditions with other instances of the test suite.