Reduce the amount of seed characteristics we test for
As our spreadsheet changes this code is fragile. Instead of requiring a full definition of our expected characteristics to be defined in our tests, we use a sample that should be easier to manage.

Ideally we'd spend more time refactoring how our characteristics are managed to make these changes easier to make. Perhaps by defining them in a yaml file that can be included where we need it.