Update seeding of bedspace characteristics
This keeps it in line with v4 of the spreadsheet.

As with premises, we reduce the amount of characteristics we define in our test to make them less fragile. If we need to rename or change in future, we won't have to update the definition in multiple places: the test and the seed task.