Update placementCriteria list
Some of these criteria are not needed for matching purposes. And there
are some that were not already included. I’ve also  added the AP types
as criteria, because they are not currently being stored in the API
properly. We may need to retire the apType field, but I want to address
one thing at a time!