Add table to store Notify Guest List users
Notify has a "Guest List" mode which allows actually sending emails in non-prod environments to specific accounts.  However, the API key is different to the normal test API key which just responds as though an email has been sent.  To allow us sending emails to the product team in dev but not erroring on other emails we want to switch between the two API keys depending on the user - this table will allow us to do this.