Update Assessment Acceptance endpoint
This updates the `assessment/acceptance` endpoint to accept two new
arguments, a `placementDates` object with the start date and placement
length, and a `notes` field. If the `placementDates` field is blank, we
create placement requirements, but not a placement request. In future
PRs, we will use these requiements to allow PPs to create Ad-hoc
placement requests.

I’ve also broken the AssessmentService tests up a bit to make them
easier to refactor / review, and DRYed it up a bit. Hope this is OK!