Update Open API spec for CAS2 applications
Adds CAS2 as service name, as well as CAS2 application models.

This requires various changes where we are checking for service names in the codebase, including extending getAllApplicationsForUsername to accommodate CAS2 applications.

As CAS2 will not handle bookings, we have to ensure CAS2 cannot be generated as a service name during booking tests.