Add an endpoint to create a placement application
This adds an endpoint to create a placement application for an
approved application.

I’ve also added some helpers to extract the entity from the
Validatable and Authorisable action results, as it’s a thing
we do a lot. Time constraints have meant I haven’t had time
to replace all the uses with the helper, but it’d be good if
we could when/if we have time!