Run tests in CircleCI
This moves the running of the tests to CircleCI with seperate
build, lint and test jobs. The build step builds the code, and
persists the built code to a workspace, which is then used by
the lint and test jobs. The test job uses parellelism to split the
tests and run them over four seperate containers, speeding up the build.
On initial inspection, this seems to shave at least 5 minutes off the
total build time!