Make user optional for Booking methods
The user is only used for role checking and emitting the domain event.  When seeding existing Bookings from a CSV we're guaranteed to be using an OfflineApplication which we can't emit Domain Events for anyway.  As we don't know which user is making the request due to lack of a JWT we can't provide one, so we will instead pass null to ignore the role checks.