Update the tasks endpoint to get tasks for a user
This moves the `/tasks` endpoint to be `/tasks/reallocatable`, which
fetches all reallocatable tasks for a workflow manager. Hitting `/tasks`
now gets all tasks for a logged in user. At the moment, this only works
for matchers, but we can easily extend this to return assessments, and
any other type of tasks we choose to add in the future.

I’ve also updated some of our `Given` functions to allow us to use them
inline, so we avoid a huge amount of nesting when we’re setting up a
load of state.