Attempt to fix CI memory errors
We’re still getting intermittent failures with the message:

```
Gradle build daemon disappeared unexpectedly (it may have been killed or may have crashed)
```

Even though we’ve reduced the memory usage, there’s still no room for
the other processes in the container sometimes, and this causes the
containers to run out of memory. This reduces the amount of RAM each
Gradle process can use, as well as upping the number of containers to
try and make up for the slower tests.