Remove Notify org.json exclusion & add workaround for duplicate keys
The way we are using OpenAPI currently results in duplicate keys being output for models with a discriminator.  We have to continue to output models this way (with both a discriminator and a normal property with the same name) as this is required by the frontend generator (and is valid OpenAPI.)  This replaces the Jackson annotation which is causing this setup to output duplicate keys in response bodies with one that does not.