Update seeding process to output root exception message too where available
Whilst the full exceptions are logged too it's not practical to go through hundreds of entries in AppInsights.  The particular example here is seeding users - the exception message doesn't identify why we failed to retrieve a user but the root exception does (e.g. a 404 vs a failure to map probation region).