Update the API spec
This adds a new typed `NewArrival` type, which, for CAS1
allows an `arrivalDateTime` field to be passed in, as well
as adding an `arrivalTime` field to the `Arrival` object.
For CAS2 and CAS3 this will always be midnight, unless they
choose to record times of arrival in the future. I decided to
have the arrival time as a separate field to allow for better
backward compatibility as updating the return type to a
DateTime might have unexpected results, especially given
that they’re just typed as strings in Typescript, so won’t
trigger a type update.