Hide bookings for LAO offenders from bedspace search endpoint when user does not have LAO access
Currently, if a user does not have LAO access, when they search for bookings the API will return
a `403 Forbidden` response if one of the results is for a LAO offender.
This will have the effect of logging the user out of the application on the frontend, making the
booking search inaccessible.

Instead, these bookings should be filtered out when the API computes the response so that the
user can successfully see bookings for which they have permissions.