Allow overbookings for AP when extending
The /extensions endpoint currently returns a 409 when there are
conflicts with an AP Booking / Lost Bed. As in other endpoints, we
accept overbookings, as this is how APs are used to working.

I’ve also removed some duplicated checks in the controller for
TA premises, as these are already taken care of in the service
method.