Filter out Documents without IDs
A small amount of documents don’t seem to have IDs, as we treat them as
non nullable, when we get them back from the API, an error is thrown.
We need to IDs to be able to download the documents, so it’s not as
simple as making the ID nullable. To get around this for now, we’ll
filter out any documnets that have no ID, so they don’t get returned.
This will unblock the immediate issue that some applicants have. If it
transpires that users are expecting those documents without IDs to be
visible, we’ll revisit.