Remove cas3 user role seeding in dev and test
We’ve found that this mechanism of seeding isn’t working. Each time the API deploys to dev the CAS3 roles are not set.

When we seed them manually they are set.
When we destroy a single pod they are set.

We think that the fact that a deployment launches multiple pods at once that a race condition means that users lost their roles. This is unconfirmed but we think the right way to fix this is to use Redlock to prevent the race.

For now we need user roles to be stable. We can’t reasonably reseed after every API deploy so we’re going to remove this and trust in manual seeding.