add UpdateApplicationType enum
We would like to be able to update a CAS2 application. The options for
the `type` field should only ever be `CAS1`, `CAS2`, `CAS3`, because we
do not handle an `offline` type of Application. This commit adds an
enum to formalise that, and make it easier for the CAS2 frontend to know
what to pass through.