Delete booking script removes any turnaround records
When we tried to use this script on prod we get the following error[1]:

```
ERROR: update or delete on table "bookings" violates foreign key constraint "turnarounds_booking_id_fkey" on table "turnarounds"
  Detail: Key (id)=(76abf8b6-2fff-40f0-aea1-c54f4d30ff5d) is still referenced from table "turnarounds".
```

This could be because turnarounds was added after this script was introduced.

This change attempts to also clear any associated turnaround records as part of the deletion.

[1] https://ministryofjustice.sentry.io/issues/4321107098/?alert_rule_id=12408391&alert_type=issue&project=4503931792392192&referrer=slack