Allow a Booking to be created without a bedID
This updates the `placement-requests/$id/bookings` endpoint to accept
either a bedId or premisesId. As we can’t easily do an either/or null
in the OpenAPI spec, we check for the presence of either bedId or
premisesId, and raise an error if neither is present.

Bed IDs are already optional, so this shouldn’t cause any breakages
elsewhere. I’ve had a quick look around and everything seems to work
as expected when a Bed ID is missing.