Allow filtering by parole status
This adds a filter option to allow us to filter the dashboard response
by if an Placement Request is a Parole request or not. In the past we’ve
done this filtering in the frontend, but as we’re going to be moving
towards pagination soon, it makes sense to do this filtering on the API
side.