add status to CAS2 ApplicationSummaries
We would like to have a dahsboard for CAS2 users that shows their in
progress applications.

This commit adds the `status` field to the `Cas2ApplicationSummary` and
the logic to return an `inProgress` status from the
`ApplicationsTransformer`. We don't currently have any other statuses,
but we will shortly be adding a submitted status.