Ensure getCalendarInfo search is inclusive
We’ve noticed an issue where if a booking/lost bed finishes on the first
day of the search period, it won’t show up. I’ve looked into this, and
it seems that the Postgres `OVERLAPS` functionality is not inclusive.
After a bit of Googling, it seems that using `tsrange` and the `&&`
overlaps operator seems to do the trick.

I’ve also added a test to demonstrate and fix the regression.