Add pagination to `getAllActive`
If a page is provided, we send a pageable object to the repository
method and return a `PaginationMetadata` data class with information
about what page the requester is on, as well as the size of the dataset
and the number of pages. This data class also has a convenience method
to cast the class as a header object for use in the next commit.