Seed CAS3 bookings for KSS in dev & test
KSS = Kent, Surrey & Sussex.

For one of the seeded properties, take the first bedspace and seed a booking in each of the different states.

We use seeded CRN’s for each booking that have been seeded into dev for us[1]. These CRN’s allow us to capture people in differing states such as multiple/single offence.

We name the seed file as “cas3” and “kss” since the seeding is so verbose and hard to manage. We have one for “eofe” which will allow our pen testers check if they are prevented from accessing content across region.

[1] https://dsdmoj.atlassian.net/wiki/spaces/AP/pages/4467949648/CRNs+for+Testing+in+Development+Environment