call `submitCas2Application` in `ApplicationsController`
We would like to be able to submit CAS2 Applications, in a similar
manner as CAS1 and 3 do at the moment, but with no domain events or
further actions taken - the application will just be marked with the
status 'submitted'.

This commit adds the controller to call the relevant service method.