Allow transactions to be configurable for Migration Jobs
Allow each migration job to define if it needs to be run in a transaction or not.  This is important for the next migration job to be added (updating name for all applications from upstream) as this could take a long time to complete and does not need to be in a single transaction.