Fix 403 Forbidden from being returned when a user with `CAS3_ASSESSOR` role tries to access assessments
Currently, the API returns the error response when a user tries to access an assessment that was not
created by them.
However, any assessments within the same region should be accessible to a user, given that they have the
`CAS3_ASSESSOR` role.
This fixes the authorisation logic to ensure that an assessment can be accessed by a user with this role.