add `submitted` status to CAS2 `ApplicationsTransformer`
We would like to be able to submit CAS2 Applications, in a similar
manner as CAS1 and 3 do at the moment, but with no domain events or
further actions taken - the application will just be marked with the
status 'submitted'.

This adds logic to set the status of an application when returning it to
the frontend.