Refactor `PersonUtils.getPersonDetailsForCrn`
Temporary Accommodation needs to filter out assessments for people
considered as LAO (Limited Access Offenders).
If a person is considered to be an LAO, then
`PersonUtils.getPersonDetailsForCrn` throws a `ForbiddenProblem`.
To implement the filtering of such assessments, it would be necessary
to either catch each `ForbiddenProblem` (expensive due to stack trace
generation) or to provide some other mechanism to retrieve this data.

This refactoring is the latter.
It creates a `tryGetPersonDetailsForCrn` method, which performs all
of the logic of the original method, but propagates the
`AuthorisableActionResult` wrapper to capture not found or unauthorised
scenarios, so that the caller can handle it.
The original method delegates to this new method to provide the
original behaviour to avoid making a larger scale refactoring.